+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 # PEST IntelliJ Changelog
 
 ## [Unreleased]
+### Fixed
+- Hide snapshot icon for import statements
+
 ### Added
-- Add support for `__DIR__` in `uses->in` calls 
+- Add support for `__DIR__` in `uses->in` calls
 
 ## [1.2.1]
 ### Fixed